+++ Dockerfile-client
@@ -14,6 +14,7 @@ RUN apt-get update && apt-get upgrade -y
 # copy whole project to your docker home directory.
 ADD client/ $DockerHOME/
 COPY protocol $DockerHOME/protocol
+COPY example_protocol $DockerHOME/example_protocol
 
 # install dependencies
 RUN pip install --upgrade pip
@@ -22,4 +23,4 @@ RUN pip install -r requirements.txt
 
 #CMD while true; do time.sleep 10; done
 #CMD python main.py
-CMD python -m pytest -vvx test/
+CMD python -m pytest -v test/
