+++ example_protocol/example_worker.py
@@ -3,7 +3,8 @@ import multiprocessing
 import socket
 import uuid
 
-from protocol import mrequests, mresponses
+from .prequests import UUIDRequest
+from .presponses import UUIDResponse
 from server.server_worker.worker import BaseWorker
 
 
@@ -18,17 +19,9 @@ class Worker(BaseWorker):
         BaseWorker.__init__(self, connection, address, server_run, timeout_s)
 
     def process(self, message_t, message) -> bool:
-        """
-        Overload this function to process custom messages. It should have the following:
-            - message_t: expected to be the request type
-            - message: expected to be message data
-            - RETURN: boolean, true if processed and
-                      false if message_t was not processed by custom messages.
-        """
-        if message_t == mrequests.UUIDRequest:
-            logging.info("UUID Request: %s", message)
+        if message_t == UUIDRequest:
             self._send(
-                mresponses.UUIDResponse(
+                UUIDResponse(
                     title=message.data.title, uuid=uuid.uuid4().hex
                 ).to_bytes()
             )
