+++ protocol/mresponses.py
@@ -43,25 +43,3 @@ class ShutdownResponse(Base):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.data = None
-
-
-class UUIDResponse(Base):
-    name = "uuid-res"
-
-    class UUIDData:
-        title: str
-        uuid: str
-
-        def __init__(self, title: str = None, uuid: str = None):
-            self.title = title
-            self.uuid = uuid
-
-        def __repr__(self):
-            ret = {"title": self.title, "uuid": self.uuid}
-            return f"{ret}"
-
-    data: UUIDData
-
-    def __init__(self, title: str = None, uuid: str = None, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.data = UUIDResponse.UUIDData(title=title, uuid=uuid)
